backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.7
dnspython==2.6.1
PyJWT==2.9.0
pymongo==4.10.1
pytz==2025.2
sqlparse==0.2.4
typing-extensions==4.13.2
tzdata==2025.2
